+++ cmd/installer/root_pack_update_test.go
@@ -4,12 +4,14 @@
 package installer_test
 
 import (
+	"os"
 	"path/filepath"
 	"testing"
 
 	errs "github.com/open-cmsis-pack/cpackget/cmd/errors"
 	"github.com/open-cmsis-pack/cpackget/cmd/installer"
 	"github.com/open-cmsis-pack/cpackget/cmd/utils"
+	"github.com/open-cmsis-pack/cpackget/cmd/xml"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -142,6 +144,67 @@ func TestUpdatePack(t *testing.T) {
 
 		// Make sure pack.idx never got touched
 		assert.False(utils.FileExists(installer.Installation.PackIdx))
+	})
+
+	t.Run("test updating all installed packs", func(t *testing.T) {
+		localTestingDir := "test-update-all-installed-packs"
+		assert.Nil(installer.SetPackRoot(localTestingDir, CreatePackRoot))
+		installer.UnlockPackRoot()
+		assert.Nil(installer.ReadIndexFiles())
+		defer removePackRoot(localTestingDir)
+
+		// packPath := publicLocalPack123
+		// err := installer.AddPack(packPath, !CheckEula, !ExtractEula, ForceReinstall, !NoRequirements, true, Timeout)
+		// assert.Nil(err)
+
+		// packPath := packToUpdate
+		// addPack(t, packPath, ConfigType{})
+		// removePack(t, packPath, true, true, false)
+		// packPath = filepath.Join(installer.Installation.DownloadDir, packToUpdateFileName)
+		// err := installer.AddPack(packPath, !CheckEula, !ExtractEula, !ForceReinstall, !NoRequirements, true, Timeout)
+		// assert.Nil(err)
+
+		// Inject pdsc into .Web folder
+		packPdscFilePath := filepath.Join(installer.Installation.WebDir, filepath.Base(pdscPublicLocalPack))
+		assert.Nil(utils.CopyFile(pdscPublicLocalPack, packPdscFilePath))
+
+		packPdscTag := xml.PdscTag{
+			Vendor:  "TheVendor",
+			Name:    "PublicLocalPack",
+			Version: "1.2.2",
+		}
+		assert.Nil(installer.Installation.PublicIndexXML.AddPdsc(packPdscTag))
+		assert.Nil(installer.Installation.PublicIndexXML.Write())
+
+		// Pre-install 1.2.2
+		addPack(t, publicLocalPack122, ConfigType{
+			IsPublic: true,
+		})
+
+		// Prepare URLs for downloading pack 1.2.4
+		pack124 := installer.PackType{}
+		pack124.Vendor = "TheVendor"
+		pack124.Name = "PublicLocalPack"
+		pack124.Version = "1.2.4"
+		pack124.IsPublic = true
+
+		// Prep server
+		pack124Content, err := os.ReadFile(publicLocalPack124)
+		assert.Nil(err)
+		server := NewServer()
+		server.AddRoute(pack124.PackFileName(), pack124Content)
+
+		// Inject URL into pdsc
+		pdscXML := xml.NewPdscXML(packPdscFilePath)
+		utils.UnsetReadOnly(packPdscFilePath)
+		assert.Nil(pdscXML.Read())
+		pdscXML.URL = server.URL()
+		assert.Nil(utils.WriteXML(packPdscFilePath, pdscXML))
+
+		err = installer.UpdatePack("", !CheckEula, !NoRequirements, SubCall, true, Timeout)
+
+		// Sanity check
+		assert.Nil(err)
 	})
 
 }
